transferred logic from HomePage to MonthStatsTable
