+++ src/components/MonthStatsTable/MonthStatsTable.jsx
@@ -4,17 +4,42 @@ import { HiOutlineChevronRight } from "react-icons/hi";
 import { useEffect, useState } from "react";
 import { MonthOneDay } from "./MonthOneDay/MonthOneDay";
 import { capitalizeFirstLetter } from "../../Utilits/capitalize";
-import { useSelector } from "react-redux";
 import { selectUser } from "../../redux/auth/selectors";
+import { fetchTodayWater } from "../../redux/todayWaterList/operations.js";
+import { useDispatch, useSelector } from "react-redux";
+import { fetchMonthWater } from "../../redux/monthWaterList/operations.js";
+import { getMonthName } from "../../Utilits/getMonth.js";
+import { selectMonthWaterList } from "../../redux/monthWaterList/selector.js";
+import { selectTodayWaterList } from "../../redux/todayWaterList/selectors.js";
 
-export const MonthStatsTable = ({
-  monthData,
-  month,
-  year,
-  setMonthNumber,
-  monthNumber,
-  setYear,
-}) => {
+export const MonthStatsTable = () => {
+
+    const dispatch = useDispatch();
+    const [year, setYear] = useState(2024);
+    const [monthNumber, setMonthNumber] = useState(11);
+    
+    const month = getMonthName(monthNumber);
+    const monthData = useSelector(selectMonthWaterList);
+    const waterList = useSelector(selectTodayWaterList);
+  
+    useEffect(() => {
+      dispatch(fetchTodayWater());
+    }, [dispatch]);
+  
+    const getMonthData = (year, month) => {
+      dispatch(
+        fetchMonthWater({
+          year,
+          month,
+        })
+      );
+    };
+  
+    useEffect(() => {
+      getMonthData(year, month);
+    }, [monthNumber, waterList]);
+  
+
   const user = useSelector(selectUser);
   const currentMonth = new Date().getMonth();
   const currentYear = new Date().getFullYear();
